add namespace isolation for the engine and fix test routing
