+++ CHANGELOG
@@ -1,7 +1,9 @@
 Unreleased
 ----------
 * rename WebhooksController module to WebhookVerification
-* added a webhooks_controller which automatically delegates to jobs
+* added a WebhooksController which automatically delegates to jobs
+* moved all engine controllers under the ShopifyApp namespace
+* rename SessionsController module to SessionsConcern
 
 6.4.2
 -----
