+++ README.md
@@ -6,7 +6,7 @@ Shopify Application Rails engine and generator
 
 Description
 -----------
-This gem includes some common code and generators for writing Rails applications using the Shopify API.
+This gem includes a Rails Engine and generators for writing Rails applications using the Shopify API. The Engine provides a SessionsController and all the required code for authenticating with a shop.
 
 The [example](https://github.com/Shopify/shopify_app/tree/master/example) directory contains an app that was generated with this gem. It also contains sample code demonstrating the usage of the embedded app sdk.
 
@@ -89,7 +89,7 @@ Other options include:
 
 You can update any of these settings later on easily, the arguments are simply for convenience.
 
-The generator creates a basic SessionsController for authenticating with your shop and a HomeController which displays basic information about your products using the ShopifyAPI. The generated controllers include concerns provided by this gem - in this way code sharing is possible and if some of these core methods are updated everyone can benefit. It is completely safe to override any of the methods provided by this gem in your application.
+The generator adds ShopifyApp and the required initializers to the host Rails application.
 
 After running the `install` generator, you can start your app with `bundle exec rails server` and install your app by visiting localhost.
 
@@ -180,7 +180,7 @@ If you only run the install generator then by default you will have an in memory
 AuthenticatedController
 -----------------------
 
-The engine includes a controller called `AuthenticatedController` which inherits from `ApplicationController`. It adds some before_filters which ensure the user is authenticated and will redirect to the login page if not. It is best practice to have all controllers that belong to the Shopify part of your app inherit from this controller. The HomeController that is generated already inherits from AuthenticatedController.
+The engine includes a controller called `ShopifyApp::AuthenticatedController` which inherits from `ApplicationController`. It adds some before_filters which ensure the user is authenticated and will redirect to the login page if not. It is best practice to have all controllers that belong to the Shopify part of your app inherit from this controller. The HomeController that is generated already inherits from AuthenticatedController.
 
 Troubleshooting
 ---------------
