+++ Rakefile
@@ -1,10 +1,6 @@
 require 'bundler/gem_tasks'
 require 'rake/testtask'
 
-Rake::TestTask.new do |t|
-  t.libs << 'test'
-  t.test_files = FileList['test/**/*_test.rb']
-  t.verbose = true
-end
+require File.expand_path('../test/dummy/config/application', __FILE__)
 
-task :default => :test
+Rails.application.load_tasks
