--- app/controllers/sessions_controller.rb
@@ -1,3 +1,0 @@
-class SessionsController < ApplicationController
-  include ShopifyApp::SessionsController
-end
