+++ app/controllers/shopify_app/sessions_controller.rb
@@ -1,0 +1,5 @@
+module ShopifyApp
+  class SessionsController < ApplicationController
+    include ShopifyApp::SessionsConcern
+  end
+end
