+++ app/views/shopify_app/sessions/new.html.erb
