+++ lib/generators/shopify_app/home_controller/templates/home_controller.rb
@@ -1,4 +1,4 @@
-class HomeController < AuthenticatedController
+class HomeController < ShopifyApp::AuthenticatedController
   def index
     @products = ShopifyAPI::Product.find(:all, :params => {:limit => 10})
   end
