+++ lib/shopify_app.rb
@@ -17,7 +17,7 @@ require 'shopify_app/webhooks_manager_job'
 require 'shopify_app/shop'
 require 'shopify_app/session_storage'
 require 'shopify_app/controller'
-require 'shopify_app/sessions_controller'
+require 'shopify_app/sessions_concern'
 require 'shopify_app/login_protection'
 require 'shopify_app/webhooks_manager'
 require 'shopify_app/webhook_verification'
