+++ lib/shopify_app/engine.rb
@@ -1,5 +1,6 @@
 module ShopifyApp
   class Engine < Rails::Engine
     engine_name 'shopify_app'
+    isolate_namespace ShopifyApp
   end
 end
