+++ test/routes/sessions_routes_test.rb
@@ -9,19 +9,18 @@ class SessionsRoutesTest < ActionController::TestCase
   end
 
   test "login routes to sessions#new" do
-    assert_routing '/login', { controller: "sessions", action: "new" }
+    assert_routing '/login', { controller: 'shopify_app/sessions', action: "new" }
   end
 
   test "post login routes to sessions#create" do
-    assert_routing({method: 'post', path: '/login'}, { controller: "sessions", action: "create" })
+    assert_routing({method: 'post', path: '/login'}, { controller: 'shopify_app/sessions', action: "create" })
   end
 
   test "auth_shopify_callback routes to sessions#callback" do
-    assert_routing '/auth/shopify/callback', { controller: "sessions", action: "callback" }
+    assert_routing '/auth/shopify/callback', { controller: 'shopify_app/sessions', action: "callback" }
   end
 
   test "logout routes to sessions#destroy" do
-    assert_routing '/logout', { controller: "sessions", action: "destroy" }
+    assert_routing '/logout', { controller: 'shopify_app/sessions', action: "destroy" }
   end
-
 end
